Add user links file, implement logout link functionality
