+++ src/components/NavLinks.js
@@ -1,10 +1,15 @@
 import React from 'react';
 import { NavLink } from 'react-router-dom';
-import { connect } from 'react-redux';
+import UserLinks from './UserLinks';
 
-const NavLinks = ({status, user}) => {
+const NavLinks = () => {
    return (
       <nav className="text-center bg-blue-500 text-green-100 p-4">
+         <NavLink
+            className="inline-block px-4 py-2"
+            exact to="/">
+            Home
+         </NavLink>
          <NavLink
             className="inline-block px-4 py-2"
             exact to="/stocks">
@@ -15,38 +20,9 @@ const NavLinks = ({status, user}) => {
             exact to="/search">
             Search a Stock
          </NavLink>
-         <NavLink
-            className="inline-block px-4 py-2"
-            exact to="/signup">
-            Signup
-         </NavLink>
-         <NavLink
-            className="inline-block px-4 py-2"
-            exact to="/login">
-            Login
-         </NavLink>
-         {console.log(status)}
-         {status === "loggedIn" ?
-            <NavLink
-               className="inline-block px-4 py-2"
-               exact to={`/profile/${user.id}`}>
-               Profile
-            </NavLink> : null
-         }
-         <NavLink
-            className="inline-block px-4 py-2"
-            exact to="/">
-            Home
-         </NavLink>
+         <UserLinks />
       </nav>
    )
 }
 
-const mapStateToProps = state => {
-   return {
-      status: state.login.status,
-      user: state.login.user      
-   }
-}
-
-export default connect(mapStateToProps)(NavLinks);
+export default NavLinks;
