+++ src/components/Routes.js
@@ -1,10 +1,12 @@
 import React from 'react';
 import StocksContainer from '../containers/StocksContainer';
 import StockSearchForm from '../components/StockSearchForm';
+import ViewStockChart from '../components/ViewStockChart';
 import SignupForm from '../components/SignupForm';
 import LoginForm from '../components/LoginForm';
 import UserProfileContainer from '../containers/UserProfileContainer';
-import ViewStockChart from '../components/ViewStockChart';
+import LogoutContainer from '../containers/LogoutContainer';
+
 import {
    Route
  } from "react-router-dom";
@@ -19,6 +21,9 @@ export const Routes = () => {
          <Route path="/login">
             <LoginForm />
          </Route>  
+         <Route path="/logout">
+            {/* <LogoutContainer /> */}
+         </Route>
          <Route path="/profile/:id">
             <UserProfileContainer />
          </Route>
